Make Rust stable friendly by disabling bench by default
