+++ Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 authors = ["hatoo <hato2000@gmail.com>"]
 edition = "2018"
 
+[lib]
+bench = false
+
 [dependencies]
 
 [dev-dependencies]
