+++ src/lib.rs
@@ -1,7 +1,4 @@
-#![feature(test)]
 #![allow(clippy::suspicious_op_assign_impl)]
-
-extern crate test;
 
 const TRUE: &bool = &true;
 const FALSE: &bool = &false;
@@ -442,46 +439,4 @@ fn test_bitset_chomp() {
     assert_eq!((&set1 | &set2).count_ones(), 4);
     assert_eq!((&set1 & &set2).count_ones(), 2);
     assert_eq!((&set1 ^ &set2).count_ones(), 2);
-}
-
-#[bench]
-fn bench_bitset_dp(b: &mut test::Bencher) {
-    use rand::prelude::*;
-    let size = 1000;
-    let mut v = Vec::new();
-    let mut rng = StdRng::seed_from_u64(114514);
-
-    for _ in 0..size {
-        v.push(rng.next_u32() as usize % size);
-    }
-
-    b.iter(|| {
-        let mut bset = BitSet::new(100000);
-        bset.set(0, true);
-
-        for &x in &v {
-            bset |= &(&bset << x);
-        }
-    });
-}
-
-#[bench]
-fn bench_bitset_dp_shl_or(b: &mut test::Bencher) {
-    use rand::prelude::*;
-    let size = 1000;
-    let mut v = Vec::new();
-    let mut rng = StdRng::seed_from_u64(114514);
-
-    for _ in 0..size {
-        v.push(rng.next_u32() as usize % size);
-    }
-
-    b.iter(|| {
-        let mut bset = BitSet::new(100000);
-        bset.set(0, true);
-
-        for &x in &v {
-            bset.shl_or(x);
-        }
-    });
 }
